Some ReSharper linting

+++ AvaloniaMVVM/AvaloniaMVVM/Kernels/CorrelationKernels.cs
@@ -1,7 +1,4 @@
 ﻿using ILGPU;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace AvaloniaMVVM.Kernels
 {
@@ -23,9 +20,9 @@ namespace AvaloniaMVVM.Kernels
             var corrX = (float)XMath.Abs(Kernels.PearsonCorrelation(bufIn, new Index2(index.X - 1, index.Y), new Index2(index.X + 1, index.Y)));
             var corrY = (float)XMath.Abs(Kernels.PearsonCorrelation(bufIn, new Index2(index.X, index.Y - 1), new Index2(index.X, index.Y + 1)));
             
-            result[index.X, index.Y, 0] = (float)((1.0f - corrX));
-            result[index.X, index.Y, 1] = (float)(XMath.Sqrt(XMath.Pow(1.0f - corrX, 2) + XMath.Pow(1.0f - corrY, 2)));
-            result[index.X, index.Y, 2] = (float)((1.0f - XMath.Min(corrX, corrY)));
+            result[index.X, index.Y, 0] = ((1.0f - corrX));
+            result[index.X, index.Y, 1] = (XMath.Sqrt(XMath.Pow(1.0f - corrX, 2) + XMath.Pow(1.0f - corrY, 2)));
+            result[index.X, index.Y, 2] = ((1.0f - XMath.Min(corrX, corrY)));
         }
 
         public static void CorrelationSobelMap(Index2 index, ArrayView3D<float> result, ArrayView3D<byte> bufIn)
@@ -44,9 +41,9 @@ namespace AvaloniaMVVM.Kernels
             var corrX = (float)XMath.Abs(Kernels.PearsonCorrelation(bufIn, new Index2(index.X - 1, index.Y), new Index2(index.X + 1, index.Y)));
             var corrY = (float)XMath.Abs(Kernels.PearsonCorrelation(bufIn, new Index2(index.X, index.Y - 1), new Index2(index.X, index.Y + 1)));
 
-            result[index.X, index.Y, 0] = (float)((1.0f - corrX));
-            result[index.X, index.Y, 1] = (float)(XMath.Sqrt(XMath.Pow(1.0f - corrX, 2) + XMath.Pow(1.0f - corrY, 2)));
-            result[index.X, index.Y, 2] = (float)((1.0f - XMath.Min(corrX, corrY)));
+            result[index.X, index.Y, 0] = ((1.0f - corrX));
+            result[index.X, index.Y, 1] = (XMath.Sqrt(XMath.Pow(1.0f - corrX, 2) + XMath.Pow(1.0f - corrY, 2)));
+            result[index.X, index.Y, 2] = ((1.0f - XMath.Min(corrX, corrY)));
         }
     }
 }
